+++ src/k8s-configuration/azext_k8s_configuration/_client_factory.py
@@ -4,12 +4,12 @@
 # --------------------------------------------------------------------------------------------
 
 
-def cf_k8sconfiguration(cli_ctx, *_):
+def cf_k8s_configuration(cli_ctx, *_):
 
     from azure.cli.core.commands.client_factory import get_mgmt_service_client
-    from azext_k8sconfiguration.vendored_sdks import SourceControlConfigurationClient
+    from azext_k8s_configuration.vendored_sdks import SourceControlConfigurationClient
     return get_mgmt_service_client(cli_ctx, SourceControlConfigurationClient)
 
 
-def cf_k8sconfiguration_operation(cli_ctx, _):
-    return cf_k8sconfiguration(cli_ctx).source_control_configurations
+def cf_k8s_configuration_operation(cli_ctx, _):
+    return cf_k8s_configuration(cli_ctx).source_control_configurations