+++ src/k8s-configuration/azext_k8s_configuration/_help.py
@@ -7,18 +7,18 @@
 from knack.help_files import helps  # pylint: disable=unused-import
 
 
-helps['k8sconfiguration'] = """
+helps['k8s-configuration'] = """
     type: group
     short-summary: Commands to manage Kubernetes configuration.
 """
 
-helps['k8sconfiguration create'] = """
+helps['k8s-configuration create'] = """
     type: command
     short-summary: Create a Kubernetes configuration.
     examples:
       - name: Create a Kubernetes configuration
         text: |-
-            az k8sconfiguration create --resource-group MyResourceGroup --cluster-name MyClusterName \\
+            az k8s-configuration create --resource-group MyResourceGroup --cluster-name MyClusterName \\
             --cluster-type connectedClusters --name MyGitConfig --operator-instance-name OperatorInst01 \\
             --operator-namespace OperatorNamespace01 --operator-type flux --operator-params "'--git-readonly'" \\
             --repository-url git://github.com/fluxHowTo/flux-get-started --enable-helm-operator  \\
@@ -27,43 +27,43 @@ helps['k8sconfiguration create'] = """
             --ssh-known-hosts '' --ssh-known-hosts-file ''
 """
 
-helps['k8sconfiguration list'] = """
+helps['k8s-configuration list'] = """
     type: command
     short-summary: List Kubernetes configurations.
     examples:
       - name: List all Kubernetes configurations of a cluster
         text: |-
-            az k8sconfiguration list --resource-group MyResourceGroup --cluster-name MyClusterName \\
+            az k8s-configuration list --resource-group MyResourceGroup --cluster-name MyClusterName \\
             --cluster-type connectedClusters
 """
 
-helps['k8sconfiguration delete'] = """
+helps['k8s-configuration delete'] = """
     type: command
     short-summary: Delete a Kubernetes configuration.
     examples:
       - name: Delete a Kubernetes configuration
         text: |-
-            az k8sconfiguration delete --resource-group MyResourceGroup --cluster-name MyClusterName \\
+            az k8s-configuration delete --resource-group MyResourceGroup --cluster-name MyClusterName \\
             --cluster-type connectedClusters --name MyConfigurationName
 """
 
-helps['k8sconfiguration show'] = """
+helps['k8s-configuration show'] = """
     type: command
     short-summary: Show details of a Kubernetes configuration.
     examples:
       - name: Show a Kubernetes configuration
         text: |-
-            az k8sconfiguration show --resource-group MyResourceGroup --cluster-name MyClusterName \\
+            az k8s-configuration show --resource-group MyResourceGroup --cluster-name MyClusterName \\
             --cluster-type connectedClusters --name MyConfigurationName
 """
 
-helps['k8sconfiguration update'] = """
+helps['k8s-configuration update'] = """
     type: command
     short-summary: Update a Kubernetes configuration.
     examples:
       - name: Update an existing Kubernetes configuration
         text: |-
-            az k8sconfiguration update --resource-group MyResourceGroup --cluster-name MyClusterName \\
+            az k8s-configuration update --resource-group MyResourceGroup --cluster-name MyClusterName \\
             --cluster-type connectedClusters --name MyConfigurationName --enable-helm-operator \\
             --repository-url git://github.com/fluxHowTo/flux-get-started --operator-params "'--git-readonly'" \\
             --helm-operator-chart-version 1.2.0 --helm-operator-params '--set helm.versions=v3'